Handle new book form no-input edge case
If the user doesn't input anything when creating a new book, the DOM
elements are created for that book regardless with empty title, author
and pages.

The solution is to make the input fields required in HTML so the
dialog doesn't close if the form is submitted with at least one empty
field, and to only create the book if all fields are non-falsey (not
empty) with a condition in javascript.